Mejora 3: soporte para coeficientes decimales

+++ Main.java
@@ -3,17 +3,17 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int a, b, c;
+        double a, b, c;
         double discriminante, x1, x2;
 
         System.out.print("a: ");
-        a = sc.nextInt();
+        a = sc.nextDouble();
 
         System.out.print("b: ");
-        b = sc.nextInt();
+        b = sc.nextDouble();
 
         System.out.print("c: ");
-        c = sc.nextInt();
+        c = sc.nextDouble();
 
         // Validación inicial
         if (a == 0) {
